Actuator demonstration
This commit adds the actuator to our main app to showcase that it works
transparently with WebFlux as well. By default only "status" and "info"
are exposed but this can be tuned globally or per endpoint.

The health endpoint has now a dedicated reactive support. This commit
showcases that with an extra check to validate our backend is up.